ETO-STORY-26: Moved authorization filters to identity configuration and added a filter for donors.
